+++ test/server/api/data-sources.spec.js
@@ -17,7 +17,7 @@ let app;
 let mongod;
 beforeAll(async () => {
     mongod = new MongoMemoryServer();
-    process.env.MONGODB_URI = await mongod.getConnectionString();
+    process.env.MONGODB_URI = await mongod.getUri();
     mongoose = require('mongoose');
     app = require('../../../server/app');
 });